requirements.txt
Flask==2.0.1
requests==2.25.1
python-dotenv==0.19.2
google-generativeai==0.3.2
spacy==3.7.2
snownlp==0.12.3

# Please run the following command to download the Chinese model for spaCy:
# python -m spacy download zh_core_web_sm

# 安全相关的依赖包
cryptography>=41.0.0  # 用于加密操作
certifi>=2023.0.0     # 用于SSL证书验证
requests>=2.31.0      # HTTP请求库
urllib3>=2.0.0        # 底层HTTP库
PyJWT>=2.4.0          # 用于JWT令牌处理
bleach>=6.0.0         # 用于HTML净化
marshmallow>=3.19.0    # 用于数据验证
flask-talisman>=1.0.0  # 用于安全头设置
flask-wtf>=1.1.1      # 用于CSRF保护
